// frontend/schwartze-frontend/src/generated-api/model/remove-subject-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Text } from './text';

/**
 * 
 * @export
 * @interface RemoveSubjectRequest
 */
export interface RemoveSubjectRequest {
    /**
     * 
     * @type {number}
     * @memberof RemoveSubjectRequest
     */
    'subject_id'?: number;
    /**
     * 
     * @type {Text}
     * @memberof RemoveSubjectRequest
     */
    'subject_text'?: Text;
    /**
     * 
     * @type {string}
     * @memberof RemoveSubjectRequest
     */
    'language'?: string;
}